Fix recent crashes in api, show, uninstall pack (#370)
* uninstall pack missing table crash

* api catch fkey error on delete

* show: catch missing subview

+++ packages/saltcorn-data/models/pack.js
@@ -127,9 +127,11 @@ const uninstall_pack = contract(
     }
     for (const tableSpec of pack.tables) {
       const table = await Table.findOne({ name: tableSpec.name });
-      const fields = await table.getFields();
-      for (const field of fields) {
-        await field.delete();
+      if (table) {
+        const fields = await table.getFields();
+        for (const field of fields) {
+          await field.delete();
+        }
       }
     }
     for (const tableSpec of pack.tables) {
